Add support for non GL backends
* Rename RenderingContext to VideoContext
* Make VideoContext an enum, enumerating all supported backends
* Put Window inside VideoContext
* display_config.backend determines what backend is used
* Direct3D backend is stubbed
