+++ examples/window.rs
@@ -24,16 +24,26 @@ impl State for Example {
     }
 
     fn update(&mut self, _delta: Duration) -> Trans {
-        let window = &mut self.context.borrow_mut().window;
-        let mut trans = Trans::None;
-        for event in window.poll_events() {
-            match event {
-                glutin::Event::KeyboardInput(_, _, Some(glutin::VirtualKeyCode::Escape)) => trans = Trans::Quit,
-                _ => ()
-            }
+        use amethyst::context::VideoContext;
+        match self.context.borrow_mut().video_context {
+            VideoContext::OpenGL { ref window, .. } => {
+                let mut trans = Trans::None;
+                for event in window.poll_events() {
+                    match event {
+                        glutin::Event::KeyboardInput(_, _, Some(glutin::VirtualKeyCode::Escape)) => trans = Trans::Quit,
+                        glutin::Event::Closed => trans = Trans::Quit,
+                        _ => (),
+                    }
+                }
+                window.swap_buffers().unwrap();
+                trans
+            },
+// #[cfg(windows)]
+            // Window::Direct3D { ref window, .. } => {
+            //     // stub
+            //     Trans::Quit
+            // },
         }
-        window.swap_buffers().unwrap();
-        trans
     }
 
     fn on_stop(&mut self) {
@@ -42,8 +52,11 @@ impl State for Example {
 }
 
 fn main() {
-    let config = amethyst::context::Config::from_file("../config/window_example_config.yml").unwrap();
-    let context = Rc::new(RefCell::new(Context::new(config)));
-    let mut game = Application::new(Example::new(context.clone()));
+    use amethyst::context::Config;
+    let config = Config::from_file("../config/window_example_config.yml").unwrap();
+    let context = Context::new(config).unwrap();
+    let context_ref = Rc::new(RefCell::new(context));
+    let example = Example::new(context_ref.clone());
+    let mut game = Application::new(example);
     game.run();
 }
